Fixed the the default profit loss calculation. Rounded the equity curve value towhole dollars so that chart will line up.

+++ Platform/TickZoomCommon/Performance/Equity.cs
@@ -171,7 +171,7 @@ namespace TickZoom.Common
 		public bool OnIntervalClose()
 		{
 			if( graphEquity) {
-				equity[0] = CurrentEquity;
+				equity[0] = (long) CurrentEquity;
 			}
 			return true;
 		}
